+++ src/components/Header/index.tsx
@@ -1,4 +1,4 @@
-import Header from "./AppHeader";
+import Header from "./Header";
 import HeaderLogo from "./HeaderLogo";
 
 export { HeaderLogo, Header };
